Replace .txt file with .py file,
Code now takes json as input and returns simplified json file as output, with format {"counter": text_body}. Depending on specific tagger requirements, we might have to simplify this further to just text with no identification, some code for this is commented out on lines 20 & 29 (Just be aware that it is appending, not overwriting the txt files, so start with a clean slate). We also may want to save  a json file with reduced metadata (author, subreddit, etc.), though in this case, we might not need to simplify the json at all.

Features added here:
-Counter for lines skipped (errors)
-Changing file extension (lines 19 & 21)
-Changed to errors="replace" due to problems with surrogateescape.
-Fixed extraction of only comment text ("body")